fix: restore previous AppsPs and rename AppsPs to AppsContainersPs
The PR #204 introduced an API change which would have require a new major release. But as stated in the Go module documentation (https://blog.golang.org/v2-go-modules):

> These kinds of changes require effort by downstream users to migrate their code to the new API, so they should not be made without careful consideration that the benefits outweigh the costs

In this situation, I don't think that the benefits outweigh the costs. Hence I'm reverting the API changes so that we don't break the API anymore.